Added List and Edit UI connecting to RESTful Fish Service
This contains the basic functions of connecting to RESTful Fish service and
displaying all records in the list-view. From that view, records can be
deleted, and the list is refreshed. The list-view has an Add button, which
goes to the edit-view, where new records can be created. Record updates
are not implemented.